refactor: break up proxy tests
